+++ src/model/nucleic-acids/index.ts
@@ -3,3 +3,4 @@ export * from './NucleotidePattern';
 export * from './NucleicAcid';
 export * from './DNA';
 export * from './RNA';
+export * from './Gene';